+++ src/__tests__/integration/ZBClient-integration.spec.ts
@@ -1,9 +1,6 @@
-import { v4 as uuid } from 'uuid'
 import { ZBClient } from '../..'
 
 process.env.ZB_NODE_LOG_LEVEL = process.env.ZB_NODE_LOG_LEVEL || 'NONE'
-
-// Force sequential execution: https://stackoverflow.com/a/55594988
 
 describe('ZBClient', () => {
 	let zbc: ZBClient
@@ -60,36 +57,6 @@ describe('ZBClient', () => {
 		)
 		expect(workflowInstance.bpmnProcessId).toBe('hello-world')
 		expect(workflowInstance.workflowInstanceKey).toBeTruthy()
-	})
-
-	it('Can start a workflow with a message', async done => {
-		const deploy = await zbc.deployWorkflow(
-			'./src/__tests__/testdata/msg-start.bpmn'
-		)
-		expect(deploy.key).toBeTruthy()
-
-		const randomId = uuid()
-
-		await zbc.publishStartMessage({
-			name: 'MSG-START_JOB',
-			timeToLive: 1000,
-			variables: {
-				testKey: randomId,
-			},
-		})
-
-		await zbc.createWorker(
-			'test2',
-			'console-log-msg',
-			async (job, complete) => {
-				complete(job.variables)
-				expect(
-					job.customHeaders.message.indexOf('Workflow') !== -1
-				).toBe(true)
-				expect(job.variables.testKey).toBe(randomId) // Makes sure the worker isn't responding to another message
-				done()
-			}
-		)
 	})
 
 	it('Can cancel a workflow', async done => {
@@ -170,17 +137,6 @@ describe('ZBClient', () => {
 			complete(job.variables)
 			done()
 		})
-	})
-
-	it('does not retry the deployment of a broken BPMN file', async () => {
-		expect.assertions(1)
-		try {
-			await zbc.deployWorkflow(
-				'./src/__tests__/testdata/broken-bpmn.bpmn'
-			)
-		} catch (e) {
-			expect(e.message.indexOf('3 INVALID_ARGUMENT:')).toBe(0)
-		}
 	})
 
 	it("does not retry to cancel a workflow instance that doesn't exist", async () => {
