+++ src/__tests__/testdata/Worker-Failure2.bpmn
@@ -1,0 +1,71 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<bpmn:definitions xmlns:bpmn="http://www.omg.org/spec/BPMN/20100524/MODEL" xmlns:bpmndi="http://www.omg.org/spec/BPMN/20100524/DI" xmlns:di="http://www.omg.org/spec/DD/20100524/DI" xmlns:dc="http://www.omg.org/spec/DD/20100524/DC" xmlns:zeebe="http://camunda.org/schema/zeebe/1.0" id="Definitions_1vwghmj" targetNamespace="http://bpmn.io/schema/bpmn" exporter="Zeebe Modeler" exporterVersion="0.6.2">
+  <bpmn:process id="worker-failure2" name="Worker Failure 2" isExecutable="true">
+    <bpmn:startEvent id="StartEvent_1" name="Start&#10;&#10;">
+      <bpmn:outgoing>SequenceFlow_0fp53hs</bpmn:outgoing>
+    </bpmn:startEvent>
+    <bpmn:serviceTask id="ServiceTask_0g6tf5f" name="Say Hello World">
+      <bpmn:extensionElements>
+        <zeebe:taskDefinition type="console-log-worker-failure-2" retries="100" />
+        <zeebe:taskHeaders>
+          <zeebe:header key="message" value="Hello World" />
+        </zeebe:taskHeaders>
+      </bpmn:extensionElements>
+      <bpmn:incoming>SequenceFlow_0fp53hs</bpmn:incoming>
+      <bpmn:outgoing>SequenceFlow_112zghv</bpmn:outgoing>
+    </bpmn:serviceTask>
+    <bpmn:sequenceFlow id="SequenceFlow_0fp53hs" sourceRef="StartEvent_1" targetRef="ServiceTask_0g6tf5f" />
+    <bpmn:endEvent id="EndEvent_16r84dr" name="End">
+      <bpmn:incoming>SequenceFlow_17rbbvu</bpmn:incoming>
+    </bpmn:endEvent>
+    <bpmn:intermediateCatchEvent id="IntermediateCatchEvent_1smhq9d" name="Wait forever">
+      <bpmn:incoming>SequenceFlow_112zghv</bpmn:incoming>
+      <bpmn:outgoing>SequenceFlow_17rbbvu</bpmn:outgoing>
+      <bpmn:messageEventDefinition messageRef="Message_0remzna" />
+    </bpmn:intermediateCatchEvent>
+    <bpmn:sequenceFlow id="SequenceFlow_112zghv" sourceRef="ServiceTask_0g6tf5f" targetRef="IntermediateCatchEvent_1smhq9d" />
+    <bpmn:sequenceFlow id="SequenceFlow_17rbbvu" sourceRef="IntermediateCatchEvent_1smhq9d" targetRef="EndEvent_16r84dr" />
+  </bpmn:process>
+  <bpmn:message id="Message_0remzna" name="Waiting For A Train That Never Arrives">
+    <bpmn:extensionElements>
+      <zeebe:subscription correlationKey="none" />
+    </bpmn:extensionElements>
+  </bpmn:message>
+  <bpmndi:BPMNDiagram id="BPMNDiagram_1">
+    <bpmndi:BPMNPlane id="BPMNPlane_1" bpmnElement="worker-failure2">
+      <bpmndi:BPMNShape id="_BPMNShape_StartEvent_2" bpmnElement="StartEvent_1">
+        <dc:Bounds x="173" y="102" width="36" height="36" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="179" y="145" width="24" height="40" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="ServiceTask_0g6tf5f_di" bpmnElement="ServiceTask_0g6tf5f">
+        <dc:Bounds x="302" y="80" width="100" height="80" />
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNEdge id="SequenceFlow_0fp53hs_di" bpmnElement="SequenceFlow_0fp53hs">
+        <di:waypoint x="209" y="120" />
+        <di:waypoint x="302" y="120" />
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNShape id="EndEvent_16r84dr_di" bpmnElement="EndEvent_16r84dr">
+        <dc:Bounds x="637" y="102" width="36" height="36" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="645" y="145" width="20" height="14" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNShape id="IntermediateCatchEvent_1smhq9d_di" bpmnElement="IntermediateCatchEvent_1smhq9d">
+        <dc:Bounds x="499" y="102" width="36" height="36" />
+        <bpmndi:BPMNLabel>
+          <dc:Bounds x="487" y="145" width="60" height="14" />
+        </bpmndi:BPMNLabel>
+      </bpmndi:BPMNShape>
+      <bpmndi:BPMNEdge id="SequenceFlow_112zghv_di" bpmnElement="SequenceFlow_112zghv">
+        <di:waypoint x="402" y="120" />
+        <di:waypoint x="499" y="120" />
+      </bpmndi:BPMNEdge>
+      <bpmndi:BPMNEdge id="SequenceFlow_17rbbvu_di" bpmnElement="SequenceFlow_17rbbvu">
+        <di:waypoint x="535" y="120" />
+        <di:waypoint x="637" y="120" />
+      </bpmndi:BPMNEdge>
+    </bpmndi:BPMNPlane>
+  </bpmndi:BPMNDiagram>
+</bpmn:definitions>
